requirements.txt

python-telegram-bot==22.5
yt-dlp
python-dotenv
aiohttp
aiofiles
